+++ packages/api-client/src/v2/blocks/scalar-auth-selector-block/helpers/auth-options.ts
@@ -10,6 +10,7 @@ export const authOptions: Record<string, AuthOption> = {
       type: 'apiKey',
       in: 'cookie',
       name: 'apiKeyCookie',
+      'x-scalar-secret-token': '',
     },
   },
   apiKeyHeader: {
@@ -18,6 +19,7 @@ export const authOptions: Record<string, AuthOption> = {
       type: 'apiKey',
       in: 'header',
       name: 'apiKeyHeader',
+      'x-scalar-secret-token': '',
     },
   },
   apiKeyQuery: {
@@ -26,6 +28,7 @@ export const authOptions: Record<string, AuthOption> = {
       type: 'apiKey',
       in: 'query',
       name: 'apiKeyQuery',
+      'x-scalar-secret-token': '',
     },
   },
   httpBasic: {
@@ -33,6 +36,9 @@ export const authOptions: Record<string, AuthOption> = {
     payload: {
       type: 'http',
       scheme: 'basic',
+      'x-scalar-secret-token': '',
+      'x-scalar-secret-username': '',
+      'x-scalar-secret-password': '',
     },
   },
   httpBearer: {
@@ -40,6 +46,9 @@ export const authOptions: Record<string, AuthOption> = {
     payload: {
       type: 'http',
       scheme: 'bearer',
+      'x-scalar-secret-token': '',
+      'x-scalar-secret-username': '',
+      'x-scalar-secret-password': '',
     },
   },
   oauth2Implicit: {
@@ -49,8 +58,9 @@ export const authOptions: Record<string, AuthOption> = {
       flows: {
         implicit: {
           authorizationUrl: '',
-          'x-scalar-client-id': '',
-          'x-scalar-redirect-uri': '',
+          refreshUrl: '',
+          'x-scalar-secret-client-id': '',
+          'x-scalar-secret-redirect-uri': '',
           'x-scalar-secret-token': '',
           scopes: {},
         },
@@ -64,11 +74,12 @@ export const authOptions: Record<string, AuthOption> = {
       flows: {
         password: {
           tokenUrl: '',
-          'x-scalar-client-id': '',
-          'x-scalar-username': '',
-          'x-scalar-password': '',
+          refreshUrl: '',
+          'x-scalar-secret-client-id': '',
+          'x-scalar-secret-username': '',
+          'x-scalar-secret-password': '',
           'x-scalar-secret-token': '',
-          'x-scalar-client-secret': '',
+          'x-scalar-secret-client-secret': '',
           scopes: {},
         },
       },
@@ -81,9 +92,10 @@ export const authOptions: Record<string, AuthOption> = {
       flows: {
         clientCredentials: {
           tokenUrl: '',
-          'x-scalar-client-id': '',
+          refreshUrl: '',
+          'x-scalar-secret-client-id': '',
           'x-scalar-secret-token': '',
-          'x-scalar-client-secret': '',
+          'x-scalar-secret-client-secret': '',
           scopes: {},
         },
       },
@@ -97,9 +109,11 @@ export const authOptions: Record<string, AuthOption> = {
         authorizationCode: {
           authorizationUrl: '',
           tokenUrl: '',
-          'x-scalar-client-id': '',
-          'x-scalar-client-secret': '',
-          'x-scalar-redirect-uri': '',
+          refreshUrl: '',
+          'x-scalar-secret-client-id': '',
+          'x-scalar-secret-client-secret': '',
+          'x-scalar-secret-token': '',
+          'x-scalar-secret-redirect-uri': '',
           'x-usePkce': 'no',
           scopes: {},
         },