+++ packages/workspace-store/src/mutators/index.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest'
 
 import { createWorkspaceStore, generateClientMutators } from '@/client'
+import type { OAuthFlowAuthorizationCode } from '@/schemas/v3.1/strict/oauth-flow'
 
 describe('generateClientMutators', () => {
   describe('cookieMutators', () => {
@@ -230,6 +231,7 @@ describe('generateClientMutators', () => {
           type: 'apiKey',
           name: 'api_key',
           in: 'header',
+          'x-scalar-secret-token': '1234567890',
         }),
       ).toBe(true)
 
@@ -238,6 +240,7 @@ describe('generateClientMutators', () => {
           type: 'apiKey',
           name: 'api_key',
           in: 'header',
+          'x-scalar-secret-token': '1234567890',
         },
       })
 
@@ -246,6 +249,7 @@ describe('generateClientMutators', () => {
           type: 'apiKey',
           name: 'api_key',
           in: 'header',
+          'x-scalar-secret-token': '1234567890',
         }),
       ).toBe(false)
 
@@ -270,9 +274,11 @@ describe('generateClientMutators', () => {
               authorizationUrl: 'https://auth.example.com/authorize',
               tokenUrl: 'https://auth.example.com/token',
               scopes: {},
-              'x-scalar-client-id': '',
-              'x-scalar-client-secret': '',
-              'x-scalar-redirect-uri': '',
+              refreshUrl: '',
+              'x-scalar-secret-token': '',
+              'x-scalar-secret-client-id': '',
+              'x-scalar-secret-client-secret': '',
+              'x-scalar-secret-redirect-uri': '',
               'x-usePkce': 'no',
             },
           },
@@ -286,12 +292,14 @@ describe('generateClientMutators', () => {
             authorizationCode: {
               authorizationUrl: 'https://auth.example.com/authorize',
               tokenUrl: 'https://auth.example.com/token',
+              refreshUrl: '',
               scopes: {},
-              'x-scalar-client-id': '',
-              'x-scalar-client-secret': '',
-              'x-scalar-redirect-uri': '',
+              'x-scalar-secret-client-id': '',
+              'x-scalar-secret-client-secret': '',
+              'x-scalar-secret-redirect-uri': '',
+              'x-scalar-secret-token': '',
               'x-usePkce': 'no',
-            },
+            } satisfies OAuthFlowAuthorizationCode,
           },
         },
       })
@@ -309,10 +317,12 @@ describe('generateClientMutators', () => {
             authorizationCode: {
               authorizationUrl: 'https://auth.example.com/authorize',
               tokenUrl: 'https://auth.example.com/token',
+              refreshUrl: '',
               scopes: {},
-              'x-scalar-client-id': '',
-              'x-scalar-client-secret': '',
-              'x-scalar-redirect-uri': '',
+              'x-scalar-secret-client-id': '',
+              'x-scalar-secret-client-secret': '',
+              'x-scalar-secret-redirect-uri': '',
+              'x-scalar-secret-token': '',
               'x-usePkce': 'no',
             },
           },
@@ -327,11 +337,13 @@ describe('generateClientMutators', () => {
               authorizationUrl: 'https://auth.example.com/authorize',
               tokenUrl: 'https://auth.example.com/token',
               scopes: {},
-              'x-scalar-client-id': '',
-              'x-scalar-client-secret': '',
-              'x-scalar-redirect-uri': '',
+              refreshUrl: '',
+              'x-scalar-secret-client-id': '',
+              'x-scalar-secret-client-secret': '',
+              'x-scalar-secret-token': '',
+              'x-scalar-secret-redirect-uri': '',
               'x-usePkce': 'no',
-            },
+            } satisfies OAuthFlowAuthorizationCode,
           },
         },
       })